+++ PowerShellGet/public/psgetfunctions/Install-Module.ps1
@@ -55,7 +55,7 @@ function Install-Module
         [Parameter()]
         [ValidateSet("CurrentUser","AllUsers")]
         [string]
-        $Scope = "AllUsers",
+        $Scope,
 
         [Parameter(ValueFromPipelineByPropertyName=$true)]
         [ValidateNotNullOrEmpty()]
@@ -91,16 +91,27 @@ function Install-Module
     {
         Get-PSGalleryApiAvailability -Repository $Repository
 
-        if(-not (Test-RunningAsElevated) -and ($Scope -ne "CurrentUser"))
+        if($Scope -eq "AllUsers" -and -not (Test-RunningAsElevated))
         {
-            # Throw an error when Install-Module is used as a non-admin user and '-Scope CurrentUser' is not specified
-            $message = $LocalizedData.InstallModuleNeedsCurrentUserScopeParameterForNonAdminUser -f @($script:programFilesModulesPath, $script:MyDocumentsModulesPath)
+            # Throw an error when Install-Module is used as a non-admin user and '-Scope AllUsers'
+            $message = $LocalizedData.InstallModuleAdminPrivilegeRequiredForAllUsersScope -f @($script:programFilesModulesPath, $script:MyDocumentsModulesPath)
 
             ThrowError -ExceptionName "System.ArgumentException" `
                         -ExceptionMessage $message `
-                        -ErrorId "InstallModuleNeedsCurrentUserScopeParameterForNonAdminUser" `
+                        -ErrorId "InstallModuleAdminPrivilegeRequiredForAllUsersScope" `
                         -CallerPSCmdlet $PSCmdlet `
                         -ErrorCategory InvalidArgument
+        }
+
+        # If no scope is specified, default installation will be to AllUsers only 
+        # If running admin on Windows with PowerShell less than v6.
+        if (-not $Scope) 
+        {
+            $Scope = "CurrentUser"
+            if(-not $script:IsCoreCLR -and (Test-RunningAsElevated))
+            {
+                $Scope = "AllUsers"
+            }
         }
 
         Install-NuGetClientBinaries -CallerPSCmdlet $PSCmdlet -Proxy $Proxy -ProxyCredential $ProxyCredential