+++ Tests/PSGetInstallModule.Tests.ps1
@@ -18,6 +18,7 @@ function SuiteSetup {
     Import-Module "$PSScriptRoot\PSGetTestUtils.psm1" -WarningAction SilentlyContinue
     Import-Module "$PSScriptRoot\Asserts.psm1" -WarningAction SilentlyContinue
     
+    $script:IsWindowsOS = (-not (Get-Variable -Name IsWindows -ErrorAction Ignore)) -or $IsWindows
     $script:ProgramFilesModulesPath = Get-AllUsersModulesPath
     $script:MyDocumentsModulesPath = Get-CurrentUserModulesPath
     $script:PSGetLocalAppDataPath = Get-PSGetLocalAppDataPath
@@ -43,7 +44,7 @@ function SuiteSetup {
     PSGetTestUtils\Uninstall-Module ContosoServer
     PSGetTestUtils\Uninstall-Module ContosoClient
 
-    if($PSEdition -ne 'Core')
+    if($script:IsWindowsOS)
     {
         $script:userName = "PSGetUser"
         $password = "Password1"
@@ -132,15 +133,19 @@ function SuiteCleanup {
     # Import the PowerShellGet provider to reload the repositories.
     $null = Import-PackageProvider -Name PowerShellGet -Force
 
-    if($PSEdition -ne 'Core')
+    if($script:IsWindowsOS)
     {
         # Delete the user
         net user $script:UserName /delete | Out-Null
         # Delete the user profile
-        $userProfile = (Get-WmiObject -Class Win32_UserProfile | Where-Object {$_.LocalPath -match $script:UserName})
-        if($userProfile)
+        # run only if cmd is available
+        if(Get-Command -Name Get-WmiObject -ErrorAction SilentlyContinue)
         {
-            RemoveItem $userProfile.LocalPath
+            $userProfile = (Get-WmiObject -Class Win32_UserProfile | Where-Object {$_.LocalPath -match $script:UserName})
+            if($userProfile)
+            {
+                RemoveItem $userProfile.LocalPath
+            }
         }
     }
       
@@ -513,18 +518,68 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         Assert ($mod2.Count -ge 2) "Atleast two versions of ContosoServer should be available after changing the -Scope with -Force and without -AllowClobber on Install-Module cmdlet, $mod2"
     }
 
-    # Purpose: InstallModuleNeedsCurrentUserScopeParameterForNonAdminUser
+    # Purpose: Install a module with all users scope parameter for non-admin user
     #
-    # Action: try to install a module without current user scope in a non-admin console
+    # Action: Try to install a module with all users scope in a non-admin console
     #
-    # Expected Result: it should fail with an error
+    # Expected Result: It should fail with an error
     #
-    It "InstallModuleNeedsCurrentUserScopeParameterForNonAdminUser" {
+    It "InstallModuleWithAllUsersScopeParameterForNonAdminUser" {
         $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
 
-        Start-Process "$PSHOME\PowerShell.exe" -ArgumentList '$null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
-                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              Install-Module -Name ContosoServer' `
+        $psProcess = "$pshome\PowerShell.exe"
+        if ($script:IsCoreCLR)
+        {
+            $psProcess = "$pshome\pwsh.exe"
+        }
+
+        Start-Process $psProcess -ArgumentList '-command if(-not (Get-PSRepository -Name PoshTest -ErrorAction SilentlyContinue)) {
+                                                Register-PSRepository -Name PoshTest -SourceLocation https://www.poshtestgallery.com/api/v2/ -InstallationPolicy Trusted
+                                                }
+                                                Install-Module -Name ContosoServer -scope AllUsers -Repository PoshTest -ErrorVariable ev -ErrorAction SilentlyContinue;
+                                                Write-Output "$ev"' `
+                                -Credential $script:credential `
+                                -Wait `
+                                -WorkingDirectory $PSHOME `
+                                -RedirectStandardOutput $NonAdminConsoleOutput
+
+        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
+        $content = Get-Content $NonAdminConsoleOutput
+        RemoveItem $NonAdminConsoleOutput
+
+        AssertNotNull ($content) "Install-Module with AllUsers scope on non-admin user console should not succeed"
+        Assert ($content -match "Administrator rights are required to install") "Install module with AllUsers scope on non-admin user console should fail, $content"
+    } `
+    -Skip:$(
+        $whoamiValue = (whoami)
+
+        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
+        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
+        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        (-not $script:IsWindowsOS) -or
+        # Temporarily disable tests for Core
+        ($script:IsCoreCLR)
+
+    )
+
+    # Purpose: Install a module with default scope parameter for non-admin user
+    #
+    # Action: Try to install a module with default (current user) scope in a non-admin console
+    #
+    # Expected Result: It should succeed and install only to current user
+    #
+    It "InstallModuleDefaultUserScopeParameterForNonAdminUser" {
+        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
+
+        $psProcess = "PowerShell.exe"
+        if ($script:IsCoreCLR)
+        {
+            $psProcess = "pwsh.exe"
+        }
+
+        Start-Process $psProcess -ArgumentList '-command Install-Module -Name ContosoServer -Repository PoshTest;
+                                                Get-InstalledModule -Name ContosoServer | Format-List Name, InstalledLocation' `
                                                -Credential $script:credential `
                                                -Wait `
                                                -WorkingDirectory $PSHOME `
@@ -533,19 +588,21 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
         $content = Get-Content $NonAdminConsoleOutput
         RemoveItem $NonAdminConsoleOutput
-        Assert ($content -match "InstallModuleNeedsCurrentUserScopeParameter") "Install module without currentuser scope on non-admin user console should fail, $content"
-        $mod = Get-Module ContosoServer -ListAvailable
-        Assert (-not $mod) "Install module without currentuser scope on non-admin user console should not install"
+
+        AssertNotNull ($content) "Install-Module with default current user scope on non-admin user console should succeed"
+        Assert ($content -match "ContosoServer") "Module did not install correctly"
+        Assert ($content -match "Documents") "Module did not install to the correct location"
     } `
     -Skip:$(
         $whoamiValue = (whoami)
 
-        ($PSEdition -eq 'Core') -or
         ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
         ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
         ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
-        ($env:APPVEYOR_TEST_PASS -eq 'True') -or
-        ($PSVersionTable.PSVersion -lt '4.0.0')
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        (-not $script:IsWindowsOS) -or
+        # Temporarily disable tests for Core
+        ($script:IsCoreCLR)
     )
 
     # Purpose: ValidateModuleIsInUseError
@@ -562,7 +619,6 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
                                                               Import-Module -Name DscTestModule;
                                                               Install-Module -Name DscTestModule -Scope CurrentUser -Force' `
                                                -Wait `
-                                               -WorkingDirectory $PSHOME `
                                                -RedirectStandardOutput $NonAdminConsoleOutput
         waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
         $content = Get-Content $NonAdminConsoleOutput
@@ -907,7 +963,7 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         $env:PSModulePath = $script:TestPSModulePath
         try {
             # Install v1 of signed module
-            Install-Module Pester -RequiredVersion 99.99.99.98 -Repository $script:localGalleryName -ErrorVariable iev -WarningVariable iwv -Force
+            Install-Module Pester -RequiredVersion 99.99.99.98 -Repository $script:localGalleryName -ErrorVariable iev -WarningVariable iwv -WarningAction SilentlyContinue -Force
             # Expect: Warning and Success
             $iev | should be $null
             $iwv | should not be $null
@@ -948,6 +1004,7 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         }
     } `
     -Skip:$((-not (Get-Module PKI -ListAvailable)) -or ([Environment]::OSVersion.Version -lt '10.0'))
+
 }
 
 Describe PowerShell.PSGet.InstallModuleTests.P1 -Tags 'P1','OuterLoop' {
@@ -1312,6 +1369,10 @@ Describe PowerShell.PSGet.InstallModuleTests.P1 -Tags 'P1','OuterLoop' {
             AssertEquals $res1.Repository $RepositoryName "PSGetItemInfo object was created with wrong repository name"
 
             $expectedInstalledLocation = Join-Path $script:ProgramFilesModulesPath -ChildPath $res1.Name
+            if($script:IsCoreCLR)
+            {
+                $expectedInstalledLocation = Join-Path -Path $script:MyDocumentsModulesPath -ChildPath $res1.Name
+            }
             if($PSVersionTable.PSVersion -ge '5.0.0')
             {
                 $expectedInstalledLocation = Join-Path -Path $expectedInstalledLocation -ChildPath $res1.Version