+++ Tests/PSGetInstallScript.Tests.ps1
@@ -18,6 +18,7 @@ function SuiteSetup {
     Import-Module "$PSScriptRoot\PSGetTestUtils.psm1" -WarningAction SilentlyContinue
     Import-Module "$PSScriptRoot\Asserts.psm1" -WarningAction SilentlyContinue
 
+    $script:IsWindowsOS = (-not (Get-Variable -Name IsWindows -ErrorAction Ignore)) -or $IsWindows
     $script:ProgramFilesScriptsPath = Get-AllUsersScriptsPath 
     $script:MyDocumentsScriptsPath = Get-CurrentUserScriptsPath 
     $script:PSGetLocalAppDataPath = Get-PSGetLocalAppDataPath
@@ -44,7 +45,7 @@ function SuiteSetup {
     Get-InstalledScript -Name Fabrikam-ServerScript -ErrorAction SilentlyContinue | Uninstall-Script -Force
     Get-InstalledScript -Name Fabrikam-ClientScript -ErrorAction SilentlyContinue | Uninstall-Script -Force
 
-    if($PSEdition -ne 'Core')
+    if($script:IsWindowsOS)
     {
         $script:userName = "PSGetUser"
         $password = "Password1"
@@ -87,15 +88,19 @@ function SuiteCleanup {
     # Import the PowerShellGet provider to reload the repositories.
     $null = Import-PackageProvider -Name PowerShellGet -Force
 
-    if($PSEdition -ne 'Core')
+    if($script:IsWindowsOS)
     {
         # Delete the user
         net user $script:UserName /delete | Out-Null
         # Delete the user profile
-        $userProfile = (Get-WmiObject -Class Win32_UserProfile | Where-Object {$_.LocalPath -match $script:UserName})
-        if($userProfile)
+        # run only if cmd is available
+        if(Get-Command -Name Get-WmiObject -ErrorAction SilentlyContinue)
         {
-            RemoveItem $userProfile.LocalPath
+            $userProfile = (Get-WmiObject -Class Win32_UserProfile | Where-Object {$_.LocalPath -match $script:UserName})
+            if($userProfile)
+            {
+                RemoveItem $userProfile.LocalPath
+            }
         }
     }
 
@@ -185,7 +190,7 @@ Describe PowerShell.PSGet.InstallScriptTests -Tags 'BVT','InnerLoop' {
         $findScriptOutput = Find-Script $scriptName
         $DateTimeBeforeInstall = Get-Date
 
-        Install-Script $scriptName
+        Install-Script $scriptName -scope CurrentUser
         $res = Get-InstalledScript $scriptName
 
         AssertEquals $res.Name $scriptName "Install-Script failed to install $scriptName, $res"
@@ -203,7 +208,7 @@ Describe PowerShell.PSGet.InstallScriptTests -Tags 'BVT','InnerLoop' {
         AssertEquals $res.Repository $findScriptOutput.Repository "Invalid Repository value in Get-InstalledScript metadata, $res"
         AssertEquals $res.RepositorySourceLocation $findScriptOutput.RepositorySourceLocation "Invalid RepositorySourceLocation value in Get-InstalledScript metadata, $res"
         AssertEquals $res.PackageManagementProvider $findScriptOutput.PackageManagementProvider "Invalid PackageManagementProvider value in Get-InstalledScript metadata, $res"
-        AssertEquals $res.InstalledLocation $script:ProgramFilesScriptsPath "Invalid InstalledLocation value in Get-InstalledScript metadata, $res"
+        AssertEquals $res.InstalledLocation $script:MyDocumentsScriptsPath "Invalid InstalledLocation value in Get-InstalledScript metadata, $res"
         AssertEquals $res.PowerShellGetFormatVersion $findScriptOutput.PowerShellGetFormatVersion "Invalid PowerShellGetFormatVersion value in Get-InstalledScript metadata, $res"
 
         AssertNotNull $res.InstalledDate "Get-InstalledScript results are not expected, InstalledDate should not be null, $res"
@@ -493,17 +498,26 @@ Describe PowerShell.PSGet.InstallScriptTests -Tags 'BVT','InnerLoop' {
         AssertEquals $res2.InstalledLocation $script:ProgramFilesScriptsPath "Install-Script with AllUsers scope and -Force did not install Fabrikam-ServerScript to program files scripts folder, $res2"
     }
 
-    # Purpose: InstallScriptNeedsCurrentUserScopeParameterForNonAdminUser
+    # Purpose: Install a script with all users scope parameter for non-admin user
     #
-    # Action: try to install a script without current user scope in a non-admin console
+    # Action: Try to install a script with all users scope in a non-admin console
     #
-    # Expected Result: it should fail with an error
+    # Expected Result: It should fail with an error
     #
-    It "InstallScriptNeedsCurrentUserScopeParameterForNonAdminUser" {
+    It "InstallScriptWithAllUsersScopeParameterForNonAdminUser" {
         $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
-        Start-Process "$PSHOME\PowerShell.exe" -ArgumentList '$null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
-                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              Install-Script -Name Fabrikam-ServerScript -Scope AllUsers' `
+
+        $psProcess = "PowerShell.exe"
+        if ($script:IsCoreCLR)
+        {
+            $psProcess = "pwsh.exe"
+        }
+
+        Start-Process $psProcess -ArgumentList '-command if(-not (Get-PSRepository -Name PoshTest -ErrorAction SilentlyContinue)) {
+                                                    Register-PSRepository -Name PoshTest -SourceLocation https://www.poshtestgallery.com/api/v2/ -InstallationPolicy Trusted
+                                                }
+                                                Install-Script -Name Fabrikam-Script -NoPathUpdate -Scope AllUsers -ErrorVariable ev -ErrorAction SilentlyContinue;
+                                                Write-Output "$ev"' `
                                                -Credential $script:credential `
                                                -Wait `
                                                -WorkingDirectory $PSHOME `
@@ -512,9 +526,9 @@ Describe PowerShell.PSGet.InstallScriptTests -Tags 'BVT','InnerLoop' {
         waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Script on non-admin console failed to complete"
         $content = Get-Content $NonAdminConsoleOutput
         RemoveItem $NonAdminConsoleOutput
-        Assert ($content -match "InstallScriptNeedsCurrentUserScopeParameter") "Install script without currentuser scope on non-admin user console should fail, $content"
-        $res = Get-InstalledScript Fabrikam-ServerScript -ErrorAction SilentlyContinue -WarningAction SilentlyContinue
-        Assert (-not $res) "Install script without currentuser scope on non-admin user console should not install, $res"
+
+        AssertNotNull ($content) "Install-Script with AllUsers scope on non-admin user console should not succeed"
+        Assert ($content -match "Administrator rights are required to install" ) "Install script with AllUsers scope on non-admin user console should fail, $content"
     } `
     -Skip:$(
         $whoamiValue = (whoami)
@@ -522,9 +536,52 @@ Describe PowerShell.PSGet.InstallScriptTests -Tags 'BVT','InnerLoop' {
         ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
         ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
         ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
-        ($env:APPVEYOR_TEST_PASS -eq 'True') -or
-        ($PSEdition -eq 'Core') -or
-        ($PSVersionTable.PSVersion -lt '4.0.0')
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        (-not $script:IsWindowsOS) -or
+        # Temporarily disable tests for Core
+        ($script:IsCoreCLR)
+    )
+
+    # Purpose: Install a script with default scope parameter for non-admin user
+    #
+    # Action: Try to install a script with default (current user) scope in a non-admin console
+    #
+    # Expected Result: It should succeed and install only to current user
+    #
+    It "InstallScriptDefaultUserScopeParameterForNonAdminUser" {
+        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
+
+        $psProcess = "PowerShell.exe"
+        if ($script:IsCoreCLR)
+        {
+            $psProcess = "pwsh.exe"
+        }
+
+        Start-Process $psProcess -ArgumentList '-command Install-Script -Name Fabrikam-ServerScript -NoPathUpdate;
+                                                Get-InstalledScript Fabrikam-ServerScript | Format-List Name, InstalledLocation' `
+                                               -Credential $script:credential `
+                                               -Wait `
+                                               -WorkingDirectory $PSHOME `
+                                               -RedirectStandardOutput $NonAdminConsoleOutput
+
+        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Script on non-admin console failed to complete"
+        $content = Get-Content $NonAdminConsoleOutput
+        RemoveItem $NonAdminConsoleOutput
+
+        AssertNotNull ($content) "Install-Script with default current user scope on non-admin user console should succeed"
+        Assert ($content -match "Fabrikam-ServerScript") "Script did not install correctly"
+        Assert ($content -match "Documents") "Script did not install to the correct location"
+    } `
+    -Skip:$(
+        $whoamiValue = (whoami)
+
+        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
+        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
+        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        (-not $script:IsWindowsOS) -or
+        # Temporarily disable tests for Core
+        ($script:IsCoreCLR)
     )
 
     # Purpose: InstallScript_AllUsers_NO_toThePromptForAddingtoPATHVariable
@@ -1714,7 +1771,12 @@ Describe PowerShell.PSGet.InstallScriptTests.P1 -Tags 'P1','OuterLoop' {
             AssertEquals $res1.RepositorySourceLocation $Global:PSGallerySourceUri "PSGetItemInfo object was created with wrong RepositorySourceLocation"
             AssertEquals $res1.Repository $RepositoryName "PSGetItemInfo object was created with wrong repository name"
 
-            AssertEquals $res1.InstalledLocation $script:ProgramFilesScriptsPath "Invalid InstalledLocation value on PSGetItemInfo object"
+            $expectedInstalledLocation = $script:ProgramFilesScriptsPath
+            if($script:IsCoreCLR)
+            {
+                $expectedInstalledLocation = $script:MyDocumentsScriptsPath
+            }
+            AssertEquals $res1.InstalledLocation $expectedInstalledLocation "Invalid InstalledLocation value on PSGetItemInfo object"
         }
         finally
         {